Fixed the beamformers with new data format
- tzeros are now consistently called initial_times. This is because in
the data format we alreadt have t0_delays which are the time instants of the
transmit firings. These in turn are called t0_delays and not tx_delays,
because in the beamforming process we already have tx_delays which are the
times needed for a transmit to reach each pixel.
- The beamformer now takes the initial_times as an input from the USBMD
dataset.
- tzero_correct has been removed from schema as this is now in the dataset
by default and is set to 0 when no correction is needed.